Add files via upload
The main file. This will create a passwords.txt file once opened and will encrypt all passwords saved to the file. The master password is "hi" but can be changed in the code. Will add an update soon for a widget to allow user to create master password within the GUI